# generated/openapi/openapi.yaml
# ─────────────────────────────────────────────────────────────────────────────
# GENERATED FILE — do not edit manually.
# Source: cms.canonical-model.yaml v1.0.0
# Generator: codegen.js
# Regenerate: node codegen.js cms.canonical-model.yaml
# ─────────────────────────────────────────────────────────────────────────────

openapi: 3.1.0
info:
  title: cms API
  version: 1.0.0
  description: cms domain — minimal starter model generated by fabric init.
paths:
  /tasks:
    post:
      operationId: createTask
      summary: Create a new task in pending status.
      tags:
        - Task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Task title is missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorError"
    get:
      operationId: listTasks
      summary: List all tasks with optional status filter.
      tags:
        - Task
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  cursor:
                    type: string
                    nullable: true
                  total:
                    type: integer
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - pending
            - in-progress
            - done
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - title
        - status
    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - pending
            - in-progress
            - done
      required:
        - title
        - status
    PaginatedCollection:
      type: object
      properties:
        data:
          type: array
          items: {}
        cursor:
          type: string
          nullable: true
        total:
          type: integer
    ApiError:
      type: object
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
      required:
        - statusCode
        - code
        - message
    ValidationErrorError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT